Update to the redirected endpoints
Commiting via GUI for first time instead of Git Bash. This data source is no longer updated by covid19india.org after 16.08.2021. Schema is changed, so massive overhaul needed to the app to continue daily updates. Therefore taking it down, until changes are made to the app to work with the new schema.